requirements.txt
Django==5.2.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
gunicorn
psycopg2-binary
dj-database-url
whitenoise
mongoengine==0.29.1